Desvincula os atores da etapa e vincula a proposição
