+++ api/management/commands/import_utils.py
@@ -242,16 +242,15 @@ def import_emendas():
 
 def import_atores():
     '''Carrega Atores'''
-    atores_df = pd.read_csv('data/atores.csv').groupby(['id_leggo', 'casa', 'id_ext'])
+    atores_df = pd.read_csv('data/atores.csv').groupby(['id_leggo'])
     for group_index in atores_df.groups:
-        prop_id = {
-            'casa': group_index[1],
-            'id_ext': group_index[2],
+        id_leggo = {
+            'id_leggo': group_index
         }
 
-        etapa_prop = get_etapa_proposicao(prop_id)
+        prop = get_proposicao(id_leggo)
 
-        if etapa_prop is None:
+        if prop is None:
             continue
 
         group_df = (
@@ -260,7 +259,7 @@ def import_atores():
             [['id_leggo', 'id_ext', 'casa', 'id_autor', 'nome_autor', 'partido', 'uf',
              'peso_total_documentos', 'tipo_generico', 'sigla_local',
               'is_important', 'bancada']]
-            .assign(proposicao=etapa_prop)
+            .assign(proposicao=prop)
         )
         Atores.objects.bulk_create(
             Atores(**r[1].to_dict()) for r in group_df.iterrows())
